Converted GBColor to a static library of functions and changed all GBColor fields to regular java.awt.Color fields.
Forgot to commit before so proceeding so this commit also includes some
portal changes.  The fine grid fades to black when zoomed all the way
out and lightens as you zoom in.  Max zoom was changed from 4 to 3
pixels per game unit (setting up to use the same portal class for the
minimap in the future).